[chore] #52 merge conflicts 해결
