+++ app/src/main/java/org/android/bbangzip/data/datasource/remote/MyPageRemoteDataSource.kt
@@ -5,7 +5,7 @@ import org.android.bbangzip.data.service.MyPageService
 import org.android.bbangzip.data.util.base.BaseResponse
 import javax.inject.Inject
 
-class MyPageDataSource @Inject constructor(
+class MyPageRemoteDataSource @Inject constructor(
     private val myPageService: MyPageService
 ){
     suspend fun getMyPageInfo(): BaseResponse<ResponseMyPageDto> = myPageService.getMyInfo()
