+++ app/src/main/java/org/android/bbangzip/di/RepositoryModule.kt
@@ -4,8 +4,7 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import org.android.bbangzip.data.repositoryImpl.MyPageRepositoryImpl
-import org.android.bbangzip.data.repositoryImpl.PieceRepositoryImpl
+import org.android.bbangzip.data.repositoryImpl.remote.PieceRepositoryImpl
 import org.android.bbangzip.data.repositoryImpl.local.UserLocalRepositoryImpl
 import org.android.bbangzip.data.repositoryImpl.remote.DummyRepositoryImpl
 import org.android.bbangzip.data.repositoryImpl.remote.MyPageRepositoryImpl
