+++ app/src/main/java/org/android/bbangzip/di/ServiceModule.kt
@@ -9,7 +9,6 @@ import org.android.bbangzip.data.service.MyPageService
 import org.android.bbangzip.data.service.PieceService
 import org.android.bbangzip.data.service.UserService
 import retrofit2.Retrofit
-import retrofit2.create
 import javax.inject.Singleton
 
 @Module
@@ -42,11 +41,4 @@ object ServiceModule {
         @BbangZip retrofit: Retrofit,
     ): UserService =
         retrofit.create(UserService::class.java)
-
-    @Provides
-    @Singleton
-    fun provideMyPageService(
-        @BbangZip retrofit: Retrofit
-    ): MyPageService =
-        retrofit.create(MyPageService::class.java)
 }
